requirements.txt
python-telegram-bot==5.3.0
Django==1.10.6
gunicorn==19.6.0
dj-database-url==0.4.1
psycopg2==2.6.2
whitenoise==3.2
urllib3==1.20
requests==2.10.0
pyTelegramBotAPI==2.1.7